struct parameters::S: unsigned can be 16-bit quantity

+++ phase0/common.h
@@ -87,7 +87,7 @@ struct parameters {
 	  * A segment is defined as a group of S consecutive blocks.
 	  * \f$ 16 \le S \le 2^{20} \f$
 	  */
-	unsigned S;
+	size_t S;
 
 	/**
 	 * \brief Subband weights for Integer DWT
